+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'counseling'
+rootProject.name = 'quiz'
