+++ src/main/java/kr/co/quiz/user/entity/Role.java
@@ -1,4 +1,4 @@
-package kr.co.counseling.user.entity;
+package kr.co.quiz.user.entity;
 
 public enum Role {
     USER, MANAGER, ADMIN
