+++ src/main/java/kr/co/quiz/user/entity/User.java
@@ -1,4 +1,4 @@
-package kr.co.counseling.user.entity;
+package kr.co.quiz.user.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
